(ref / in progress): Split static and dynamic RouteDicts. (Issue #1)
Added RouteDictGroup to group two RouteDicts (static and dynamic) together and adapted the Router to work with the group / seperate routeDicts.

However the current state is very unoptimized and produces duplicates in the seperated routeDicts. The code will be refactored (in specific GetRouteDict()), when the RoutePathType-Determination Helper (Issue #2) is finished.